Move ui handlers out of main.ts
